New booster packs and balancing
Adds a new version of the normal minor arcana pack and one jumbo version of it.

The weight of all these have been balanced to be 33% less likely to appear than the tarot packs.

Notes (not for the changelog):
Normal packs each should have a weight of 2/3, but since there are only 2 as of now, each one has 4/3, for a total weight of 8/3 between all of them, compared to a total weight of 12/3 for the normal tarot packs

Jumbo packs should each have a weight of 2/3 as well, but since there's only one, it has 4/3, for a total weight of 4/3, compared to 6/3 for jumbo tarot packs

Mega packs should have weight of 1/6, there's only one with 2/6 for a total of 2/6 (or 1/3) weight, compared to 3/6 for mega tarot packs